Separate mesh and graph related partitioning.
