+++ cpp/dolfinx/graph/CMakeLists.txt
@@ -7,6 +7,7 @@ set(HEADERS_graph
   ${CMAKE_CURRENT_SOURCE_DIR}/Graph.h
   ${CMAKE_CURRENT_SOURCE_DIR}/KaHIP.h
   ${CMAKE_CURRENT_SOURCE_DIR}/ParMETIS.h
+  ${CMAKE_CURRENT_SOURCE_DIR}/Partitioning.h
   ${CMAKE_CURRENT_SOURCE_DIR}/SCOTCH.h
   PARENT_SCOPE)
 
@@ -15,5 +16,6 @@ target_sources(dolfinx PRIVATE
   ${CMAKE_CURRENT_SOURCE_DIR}/GraphBuilder.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/KaHIP.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/ParMETIS.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/Partitioning.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/SCOTCH.cpp
 )
