+++ cpp/dolfinx/mesh/Geometry.cpp
@@ -10,6 +10,7 @@
 #include <dolfinx/common/IndexMap.h>
 #include <dolfinx/fem/DofMapBuilder.h>
 #include <dolfinx/fem/ElementDofLayout.h>
+#include <dolfinx/graph/Partitioning.h>
 #include <sstream>
 
 using namespace dolfinx;
@@ -114,7 +115,7 @@ mesh::Geometry mesh::create_geometry(
   //  processes own the cells this process needs.
   std::set<int> _src(src.begin(), src.end());
   auto [cell_nodes, global_index_cell]
-      = Partitioning::exchange(comm, cells, dest, _src);
+      = graph::Partitioning::exchange(comm, cells, dest, _src);
 
   // Build list of unique (global) node indices from adjacency list
   // (geometry nodes)
@@ -127,18 +128,18 @@ mesh::Geometry mesh::create_geometry(
   //  Fetch node coordinates by global index from other ranks. Order of
   //  coords matches order of the indices in 'indices'
   Eigen::Array<double, Eigen::Dynamic, Eigen::Dynamic, Eigen::RowMajor> coords
-      = Partitioning::distribute_data(comm, indices, x);
+      = graph::Partitioning::distribute_data(comm, indices, x);
 
   // Compute local-to-global map from local indices in dofmap to the
   // corresponding global indices in cell_nodes
   std::vector<std::int64_t> l2g
-      = Partitioning::compute_local_to_global_links(cell_nodes, dofmap);
+      = graph::Partitioning::compute_local_to_global_links(cell_nodes, dofmap);
 
   // Compute local (dof) to local (position in coords) map from (i)
   // local-to-global for dofs and (ii) local-to-global for entries in
   // coords
   std::vector<std::int32_t> l2l
-      = Partitioning::compute_local_to_local(l2g, indices);
+      = graph::Partitioning::compute_local_to_local(l2g, indices);
 
   // Build coordinate dof array
   Eigen::Array<double, Eigen::Dynamic, Eigen::Dynamic, Eigen::RowMajor> xg(
