+++ cpp/dolfinx/mesh/Topology.cpp
@@ -12,6 +12,7 @@
 #include <dolfinx/common/utils.h>
 #include <dolfinx/fem/ElementDofLayout.h>
 #include <dolfinx/graph/AdjacencyList.h>
+#include <dolfinx/graph/Partitioning.h>
 #include <numeric>
 
 using namespace dolfinx;
@@ -541,14 +542,14 @@ mesh::create_topology(MPI_Comm comm,
 
   // Distribute cells to destination rank
   const auto [my_cells, src, original_cell_index]
-      = Partitioning::distribute(comm, cells_v, dest);
+      = graph::Partitioning::distribute(comm, cells_v, dest);
 
   // Build local cell-vertex connectivity, with local vertex indices
   // [0, 1, 2, ..., n), from cell-vertex connectivity using global
   // indices and get map from global vertex indices in 'cells' to the
   // local vertex indices
   auto [cells_local, local_to_global_vertices]
-      = Partitioning::create_local_adjacency_list(my_cells);
+      = graph::Partitioning::create_local_adjacency_list(my_cells);
 
   // Create (i) local topology object and (ii) IndexMap for cells, and
   // set cell-vertex topology
@@ -595,8 +596,9 @@ mesh::create_topology(MPI_Comm comm,
   // vertices, and map from local index to old global index
   const std::vector<bool>& exterior_vertices
       = Partitioning::compute_vertex_exterior_markers(topology_local);
-  auto [cells_d, vertex_map] = Partitioning::create_distributed_adjacency_list(
-      comm, *_cells_local, local_to_global_vertices, exterior_vertices);
+  auto [cells_d, vertex_map]
+      = graph::Partitioning::create_distributed_adjacency_list(
+          comm, *_cells_local, local_to_global_vertices, exterior_vertices);
 
   Topology topology(layout.cell_type());
