+++ python/dolfinx/wrappers/mesh.cpp
@@ -289,34 +289,7 @@ void mesh(py::module& m)
       .def_static("dihedral_angles_min_max",
                   &dolfinx::mesh::MeshQuality::dihedral_angles_min_max);
 
-  // New Partition interface
-
-  m.def("create_local_adjacency_list",
-        &dolfinx::mesh::Partitioning::create_local_adjacency_list);
-  m.def("create_distributed_adjacency_list",
-        [](const MPICommWrapper comm,
-           const dolfinx::graph::AdjacencyList<std::int32_t>& list_local,
-           const std::vector<std::int64_t>& global_links,
-           const std::vector<bool>& exterior_links) {
-          return dolfinx::mesh::Partitioning::create_distributed_adjacency_list(
-              comm.get(), list_local, global_links, exterior_links);
-        });
-  m.def("distribute",
-        [](const MPICommWrapper comm,
-           const dolfinx::graph::AdjacencyList<std::int64_t>& list,
-           const dolfinx::graph::AdjacencyList<std::int32_t>& destinations) {
-          return dolfinx::mesh::Partitioning::distribute(comm.get(), list,
-                                                         destinations);
-        });
-
-  m.def("exchange",
-        [](const MPICommWrapper comm,
-           const dolfinx::graph::AdjacencyList<std::int64_t>& list,
-           const dolfinx::graph::AdjacencyList<std::int32_t>& destinations,
-           const std::set<int>& sources) {
-          return dolfinx::mesh::Partitioning::exchange(comm.get(), list,
-                                                       destinations, sources);
-        });
+  // Partitioning interface
 
   m.def("partition_cells",
         [](const MPICommWrapper comm, int nparts,
@@ -325,20 +298,6 @@ void mesh(py::module& m)
           return dolfinx::mesh::Partitioning::partition_cells(
               comm.get(), nparts, cell_type, cells);
         });
-
-  m.def("distribute_data",
-        [](const MPICommWrapper comm, const std::vector<std::int64_t>& indices,
-           const Eigen::Ref<const Eigen::Array<
-               double, Eigen::Dynamic, Eigen::Dynamic, Eigen::RowMajor>>& x) {
-          return dolfinx::mesh::Partitioning::distribute_data(comm.get(),
-                                                              indices, x);
-        });
-
-  m.def("compute_local_to_global_links",
-        &dolfinx::mesh::Partitioning::compute_local_to_global_links);
-
-  m.def("compute_local_to_local",
-        &dolfinx::mesh::Partitioning::compute_local_to_local);
 
   m.def("compute_marked_boundary_entities",
         &dolfinx::mesh::compute_marked_boundary_entities);
